refactor(runs): split open ask responsibilities (#673)
open_asks.py owned obligation creation, settlement, expiry and digest
projection in one 1104-line module. Split along the responsibility
seams: open_asks.py keeps creation and reopening, open_ask_settlement.py
takes the answer/route transitions, open_ask_digest.py takes expiry and
the straggler projection. All importers updated; no compatibility
re-exports, so each name has exactly one owner.

Closes #673
